チームの目標を設定、変更できるようにしました
