+++ lib/Screens/Home/to_teampage_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import '../../teamcreate_page.dart';
-import '../../team_page.dart';
+import '../TeamPage/team_page.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 
 class ToTeampageButton extends StatelessWidget {
